+++ app/components/map/services/raster-map-layer-service.js
@@ -78,16 +78,21 @@ angular.module('map')
       };
 
       rasterMapLayer.rescale = function (bounds) {
+        // The wms "layers" parameter should be the slug, but some raster-stores
+        // have multiple layers (i.e. rain) for different zoomlevels. Ideally
+        // we only use the uuid to identify the raster-store. This code works
+        // for elevation, but is a problematic solution.
         if (!rasterMapLayer.temporal) {
           var url = rasterMapLayer._imageUrlBase +
             '?request=getlimits&layers=' +
-            rasterMapLayer.complexWmsOptions.layers +
+            rasterMapLayer.slug +
             '&width=16&height=16&srs=epsg:4326&bbox=' +
             bounds.toBBoxString();
           $http.get(url).success(function (data) {
             var limits = ':' + data[0][0] + ':' + data[0][1];
             // strip existing domain if already present.
-            rasterMapLayer.complexWmsOptions.styles = rasterMapLayer.complexWmsOptions.styles.split(':')[0];
+            rasterMapLayer.complexWmsOptions.styles = rasterMapLayer
+            .complexWmsOptions.styles.split(':')[0];
             rasterMapLayer._imageOverlays[0].setParams({
               styles: rasterMapLayer.complexWmsOptions.styles + limits
             });