+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog of lizard-nxt client
 Unreleased (4.2.0) (XXXX-XX-XX)
 -------------------------------
 
+- Fix rescale, send raster-store slug as layers parameter to wms getLimits.
 - Undo changes to add timeseries to dashboard when switching from map.